misc/HtmlTestSuite/results/test-0126.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>Overflow: intersection observer with overflow-clip-margin</title>
<link href="https://www.w3.org/TR/css-overflow-3/#propdef-overflow-clip-margin" rel="help"/>
<link href="mailto:sky@chromium.org" rel="author" title="Scott Violet"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<style>
  #clipped_container {
    overflow: clip;
    width: 100px;
    height: 100px;
    border: solid;
    overflow-clip-margin: 50px;
  }
  #big_green_div {
      position: relative;
      width: 1000px;
      height: 1000px;
      background: green;
      left: -200px;
      top: -200px;
  }
  /* These values ensure the element is vertically offscreen. */
  .spacer { width: 150px; height: calc(100vh + 10px); }
</style>
</head><body><div class="spacer"/>
<div id="clipped_container">
  <div id="big_green_div"/>
</div>

<script>
let t = async_test("ParentWithOverflowClipMargin");
let options = {
  threshold: 0,
  rootMargin: '0px'
}
// The 'big_green_div' is initially on screen due to
// overflow-clip-margin of the parent. Once the observer is notified, the
// overflow-clip-margin is reduced so that 'big_green_div' is no longer
// on screen, and the observer should again be notified.
let gotIntersects = false;
let intersectionObserver = new IntersectionObserver((entries, observer) =&gt; {
  t.step(function() { assert_equals(1, entries.length); });
  let entry = entries[0];
  if (!gotIntersects) {
    t.step(function() { assert_true(entry.isIntersecting); });
    gotIntersects = true;
    document.getElementById('clipped_container').style.overflowClipMargin = "0px";
  } else {
    t.step(function() { assert_false(entry.isIntersecting); });
    t.done();
  }}, options);
intersectionObserver.observe(document.getElementById('big_green_div'));
</script>
</body></html>
